@@ -1,8 +1,7 @@
 name: flutter_openim_widget
-description: openim widget.
-version: 0.0.1
-homepage: https://github.com/hrxiang/OpenIM-Flutter-Widget
-publish_to: none
+description: The public ui library is used with the openim demo, and you can directly use it for secondary development.
+version: 0.0.8
+homepage: https://github.com/hrxiang/flutter_openim_widget.git
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
@@ -11,44 +10,32 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  font_awesome_flutter: ^10.0.0
   rxdart: ^0.27.3
-  flutter_screenutil: ^5.0.1
-  #  flutter_slidable: ^1.1.0
+  flutter_screenutil: ^5.3.1
+  flutter_slidable: ^1.2.0
 #  cached_network_image: ^3.2.0
-  path_provider: ^2.0.7
 #  photo_view: ^0.13.0
+  path_provider: ^2.0.9
   flutter_image_compress: ^1.1.0
   record: ^3.0.2
-  just_audio: ^0.9.18
-  #  chewie: ^1.2.2
-  #  provider: ^6.0.1
-  video_player: ^2.2.7
-  lottie: ^1.2.1
-  permission_handler: ^8.3.0
-  extended_text_field: ^9.0.1
-#  bubble: ^1.2.1
-#  bubble_box: ^0.5.1
-  focus_detector: ^2.0.0
+  just_audio: ^0.9.20
+  lottie: ^1.2.2
+  permission_handler: ^9.1.0
+  extended_text_field: ^10.0.0
+  extended_image: ^6.0.2+1
+  focus_detector: ^2.0.1
   sprintf: ^6.0.0
   flutter_inappwebview: ^5.3.2
-  animate_do: ^2.0.0
-  #  flutter_openim_sdk: ^0.0.9+4
-  #  linkable: ^3.0.1
-  #  flutter_parsed_text: ^2.2.1
-  flutter_svg: ^1.0.0
-#  transparent_image: ^2.0.0
-  extended_image: ^6.0.1
-  chewie:
-    git:
-      url: https://github.com/hrxiang/chewie.git
-  flutter_slidable:
-    git:
-      url: https://github.com/letsar/flutter_slidable.git
-  flutter_openim_sdk: ^1.0.8
-#  flutter_openim_sdk:
-#    git:
-#      url: https://github.com/OpenIMSDK/Open-IM-SDK-Flutter.git
-#      ref: fix-ios-simulator
+  animate_do: ^2.1.0
+  flutter_svg: ^1.0.3
+  qq_badge: ^0.0.2
+#  better_player: ^0.0.81
+  chewie: ^1.3.1
+  video_player: ^2.3.0
+  flutter_cache_manager: ^3.3.0
+  mime_type: ^1.0.0
+  flutter_openim_sdk: ^2.0.0+3
 #  flutter_openim_sdk:
 #    path: ../flutter_openim_sdk
 
