@@ -6,6 +6,7 @@ export 'package:flutter_slidable/flutter_slidable.dart';
 export 'src/at_special_text_span_builder.dart';
 export 'src/chat_at_text.dart';
 export 'src/chat_avatar_view.dart';
+export 'src/chat_bubble.dart';
 export 'src/chat_carte_view.dart';
 export 'src/chat_emoji_view.dart';
 export 'src/chat_file_preview.dart';
@@ -16,6 +17,7 @@ export 'src/chat_item_single_view.dart';
 export 'src/chat_itemview.dart';
 export 'src/chat_linear_progress_indicator.dart';
 export 'src/chat_location_view.dart';
+export 'src/chat_longpress_ripple.dart';
 export 'src/chat_menu.dart';
 export 'src/chat_merge_view.dart';
 export 'src/chat_multi_toolbox.dart';
@@ -36,12 +38,14 @@ export 'src/chat_voice_view.dart';
 export 'src/chat_webview_map.dart';
 export 'src/conversation_itemview.dart';
 export 'src/copy_custom_pop_up_menu.dart';
+export 'src/custom_chawie_controls.dart';
+export 'src/favorite_emoji_listview.dart';
 export 'src/pop_button.dart';
 export 'src/title_bar.dart';
 export 'src/unread_count_view.dart';
 export 'src/util/common_util.dart';
 export 'src/util/custom_ext.dart';
-export 'src/util/icon_util.dart';
+export 'src/util/image_util.dart';
 export 'src/util/permission_util.dart';
 export 'src/util/ui_locallizations.dart';
 export 'src/util/voice_record.dart';
