@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_openim_widget/flutter_openim_widget.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:rxdart/rxdart.dart';
-
-import 'at_special_text_span_builder.dart';
 
 class ChatInputBoxView extends StatefulWidget {
   ChatInputBoxView({
@@ -23,6 +22,9 @@ class ChatInputBoxView extends StatefulWidget {
     this.quoteContent,
     this.onClearQuote,
     this.multiMode = false,
+    this.inputFormatters,
+    this.showEmojiButton = true,
+    this.showToolsButton = true,
   }) : super(key: key);
   final AtTextCallback? atCallback;
   final Map<String, String> allAtMap;
@@ -39,6 +41,9 @@ class ChatInputBoxView extends StatefulWidget {
   final String? quoteContent;
   final Function()? onClearQuote;
   final bool multiMode;
+  final List<TextInputFormatter>? inputFormatters;
+  final bool showEmojiButton;
+  final bool showToolsButton;
 
   @override
   _ChatInputBoxViewState createState() => _ChatInputBoxViewState();
@@ -120,6 +125,25 @@ class _ChatInputBoxViewState extends State<ChatInputBoxView>
 
   unfocus() => FocusScope.of(context).requestFocus(FocusNode());
 
+  EdgeInsetsGeometry get emojiButtonPadding {
+    if (widget.showToolsButton) {
+      return EdgeInsets.only(left: 10.w, right: 5.w);
+    } else {
+      return EdgeInsets.only(left: 10.w, right: 10.w);
+    }
+  }
+
+  EdgeInsetsGeometry get toolsButtonPadding {
+    if (widget.showEmojiButton) {
+      return EdgeInsets.only(left: 5.w, right: 10.w);
+    } else {
+      return EdgeInsets.only(left: 10.w, right: 10.w);
+    }
+  }
+
+  SizedBox get spaceView => SizedBox(
+      width: widget.showEmojiButton || widget.showToolsButton ? 0 : 10.w);
+
   Widget _buildMsgInputField({required BuildContext context}) => Column(
         children: [
           Container(
@@ -163,12 +187,14 @@ class _ChatInputBoxViewState extends State<ChatInputBoxView>
                         ],
                       ),
                     ),
-                    _rightKeyboardButton ? _keyboardRightBtn() : _emojiBtn(),
-                    _toolsBtn(),
+                    if (widget.showEmojiButton)
+                      _rightKeyboardButton ? _keyboardRightBtn() : _emojiBtn(),
+                    if (widget.showToolsButton) _toolsBtn(),
+                    spaceView,
                     Visibility(
                       visible: !_leftKeyboardButton || !_rightKeyboardButton,
                       child: Container(
-                        width: 60.0 * (1.0 - _animation.value),
+                        width: 60.0.w * (1.0 - _animation.value),
                         child: _buildSendButton(),
                       ),
                     ),
@@ -207,6 +233,7 @@ class _ChatInputBoxViewState extends State<ChatInputBoxView>
           ),
           child: Text(
             UILocalizations.send,
+            maxLines: 1,
             style: TextStyle(
               fontSize: 14.sp,
               color: Color(0xFFFFFFFF),
@@ -237,7 +264,7 @@ class _ChatInputBoxViewState extends State<ChatInputBoxView>
                   ),
                 ),
               ),
-              IconUtil.assetImage(
+              ImageUtil.assetImage(
                 'ic_del_quote',
                 width: 14.w,
                 height: 15.h,
@@ -265,6 +292,7 @@ class _ChatInputBoxViewState extends State<ChatInputBoxView>
           allAtMap: widget.allAtMap,
           focusNode: widget.focusNode,
           controller: widget.controller,
+          inputFormatters: widget.inputFormatters,
           // onSubmitted: (value) {
           //   focus();
           //   if (null != widget.onSubmitted) widget.onSubmitted!(value);
@@ -284,22 +312,20 @@ class _ChatInputBoxViewState extends State<ChatInputBoxView>
   );
 
   Widget _speakBtn() => _buildBtn(
-        icon: IconUtil.speak(),
+        icon: ImageUtil.speak(),
         onTap: () {
-          PermissionUtil.microphone(() {
-            setState(() {
-              _leftKeyboardButton = true;
-              _rightKeyboardButton = false;
-              _toolsVisible = false;
-              _emojiVisible = false;
-              unfocus();
-            });
+          setState(() {
+            _leftKeyboardButton = true;
+            _rightKeyboardButton = false;
+            _toolsVisible = false;
+            _emojiVisible = false;
+            unfocus();
           });
         },
       );
 
   Widget _keyboardLeftBtn() => _buildBtn(
-        icon: IconUtil.keyboard(),
+        icon: ImageUtil.keyboard(),
         onTap: () {
           setState(() {
             _leftKeyboardButton = false;
@@ -311,8 +337,9 @@ class _ChatInputBoxViewState extends State<ChatInputBoxView>
       );
 
   Widget _keyboardRightBtn() => _buildBtn(
-        padding: EdgeInsets.only(left: 10.w, right: 5.w),
-        icon: IconUtil.keyboard(),
+        padding: emojiButtonPadding,
+        // padding: EdgeInsets.only(left: 10.w, right: 5.w),
+        icon: ImageUtil.keyboard(),
         onTap: () {
           setState(() {
             _rightKeyboardButton = false;
@@ -324,8 +351,9 @@ class _ChatInputBoxViewState extends State<ChatInputBoxView>
       );
 
   Widget _toolsBtn() => _buildBtn(
-        icon: IconUtil.tools(),
-        padding: EdgeInsets.only(left: 5.w, right: 10.w),
+        icon: ImageUtil.tools(),
+        // padding: EdgeInsets.only(left: 5.w, right: 10.w),
+        padding: toolsButtonPadding,
         onTap: () {
           setState(() {
             _toolsVisible = !_toolsVisible;
@@ -342,8 +370,9 @@ class _ChatInputBoxViewState extends State<ChatInputBoxView>
       );
 
   Widget _emojiBtn() => _buildBtn(
-        padding: EdgeInsets.only(left: 10.w, right: 5.w),
-        icon: IconUtil.emoji(),
+        padding: emojiButtonPadding,
+        // padding: EdgeInsets.only(left: 10.w, right: 5.w),
+        icon: ImageUtil.emoji(),
         onTap: () {
           setState(() {
             _rightKeyboardButton = true;
