@@ -1,9 +1,8 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
+import 'package:flutter_openim_widget/flutter_openim_widget.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-
-import '../flutter_openim_widget.dart';
 
 class ChatLocationView extends StatelessWidget {
   ChatLocationView({
@@ -74,7 +73,7 @@ class ChatLocationView extends StatelessWidget {
             Row(
               children: [
                 Expanded(
-                  child: IconUtil.networkImage(
+                  child: ImageUtil.networkImage(
                     url: url,
                     height: 100.h,
                     fit: BoxFit.cover,
