@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_openim_widget/flutter_openim_widget.dart';
 
-import 'chat_itemview.dart';
-
 class ChatSendFailedView extends StatefulWidget {
   final String msgId;
   final bool isReceived;
   final Stream<MsgStreamEv<bool>>? stream;
   final bool isSendFailed;
+  final Function()? onFailedResend;
 
   const ChatSendFailedView({
     Key? key,
@@ -15,6 +14,7 @@ class ChatSendFailedView extends StatefulWidget {
     required this.isReceived,
     this.isSendFailed = false,
     this.stream,
+    this.onFailedResend,
   }) : super(key: key);
 
   @override
@@ -42,7 +42,11 @@ class _ChatSendFailedViewState extends State<ChatSendFailedView> {
   Widget build(BuildContext context) {
     return Visibility(
       visible: !widget.isReceived && _failed,
-      child: IconUtil.sendFailed(),
+      child: GestureDetector(
+        behavior: HitTestBehavior.translucent,
+        onTap: widget.onFailedResend,
+        child: ImageUtil.sendFailed(),
+      ),
     );
   }
 }
