@@ -1,8 +1,8 @@
 import 'package:extended_text_field/extended_text_field.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_openim_widget/flutter_openim_widget.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-
-import 'at_special_text_span_builder.dart';
 
 class ChatTextField extends StatelessWidget {
   final AtTextCallback? atCallback;
@@ -12,6 +12,7 @@ class ChatTextField extends StatelessWidget {
   final ValueChanged<String>? onSubmitted;
   final TextStyle? style;
   final TextStyle? atStyle;
+  final List<TextInputFormatter>? inputFormatters;
 
   const ChatTextField({
     Key? key,
@@ -22,6 +23,7 @@ class ChatTextField extends StatelessWidget {
     this.onSubmitted,
     this.style,
     this.atStyle,
+    this.inputFormatters,
   }) : super(key: key);
 
   @override
@@ -50,6 +52,35 @@ class ChatTextField extends StatelessWidget {
           vertical: 8.h,
         ),
       ),
+      inputFormatters: inputFormatters,
     );
+  }
+}
+
+class AtTextInputFormatter extends TextInputFormatter {
+  final String? Function()? onTap;
+
+  AtTextInputFormatter(this.onTap);
+
+  @override
+  TextEditingValue formatEditUpdate(
+      TextEditingValue oldValue, TextEditingValue newValue) {
+    int end = newValue.selection.end;
+    int start = oldValue.selection.baseOffset;
+    if (oldValue.text.length <= newValue.text.length) {
+      var newChar = newValue.text.substring(start, end);
+      if (newChar == '@') {
+        var result = onTap?.call();
+        if (result != null) {
+          var v1 = newValue.text.replaceRange(start, end, result);
+          var offset = start + result.length;
+          return TextEditingValue(
+            text: v1,
+            selection: TextSelection.collapsed(offset: offset),
+          );
+        }
+      }
+    }
+    return newValue;
   }
 }
