@@ -1,7 +1,6 @@
 import 'package:animate_do/animate_do.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_openim_widget/flutter_openim_widget.dart';
-import 'package:flutter_openim_widget/src/chat_longpress_ripple.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 typedef IndexedToolsBuilder = Widget Function(
@@ -87,6 +86,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
     );*/
     return Container(
       height: 190.h,
+      color: Colors.white,
       child: Stack(
         children: [
           FadeInUp(
@@ -111,7 +111,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                   label: UILocalizations.album,
                   style: toolsTextStyle,
                   image: _buildBtn(
-                    icon: IconUtil.toolsAlbum(),
+                    icon: ImageUtil.toolsAlbum(),
                     onTap: widget.onTapAlbum,
                   ),
                 )),
@@ -122,7 +122,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                   label: UILocalizations.file,
                   style: toolsTextStyle,
                   image: _buildBtn(
-                    icon: IconUtil.toolsFile(),
+                    icon: ImageUtil.toolsFile(),
                     onTap: widget.onTapFile,
                   ),
                 )),
@@ -135,7 +135,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                   label: UILocalizations.camera,
                   style: toolsTextStyle,
                   image: _buildBtn(
-                    icon: IconUtil.toolsCamera(),
+                    icon: ImageUtil.toolsCamera(),
                     onTap: widget.onTapCamera,
                   ),
                 )),
@@ -146,7 +146,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                   label: UILocalizations.carte,
                   style: toolsTextStyle,
                   image: _buildBtn(
-                    icon: IconUtil.toolsCarte(),
+                    icon: ImageUtil.toolsCarte(),
                     onTap: widget.onTapCarte,
                   ),
                 )),
@@ -159,7 +159,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                   label: UILocalizations.videoCall,
                   style: toolsTextStyle,
                   image: _buildBtn(
-                    icon: IconUtil.toolsVideoCall(),
+                    icon: ImageUtil.toolsVideoCall(),
                     onTap: widget.onTapVideoCall,
                   ),
                 )),
@@ -170,7 +170,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                   label: UILocalizations.voiceInput,
                   style: toolsTextStyle,
                   image: _buildBtn(
-                    icon: IconUtil.toolsVoiceInput(),
+                    icon: ImageUtil.toolsVoiceInput(),
                     onTap: () {
                       setState(() {
                         _enabledVoiceInput = true;
@@ -186,7 +186,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
               label: UILocalizations.location,
               style: toolsTextStyle,
               image: _buildBtn(
-                icon: IconUtil.toolsLocation(),
+                icon: ImageUtil.toolsLocation(),
                 onTap: widget.onTapLocation,
               ),
             )),
@@ -208,7 +208,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                   Center(
                     child: LongPressRippleAnimation(
                       radius: 44.h,
-                      child: IconUtil.voiceInputNor(),
+                      child: ImageUtil.voiceInputNor(),
                       onStart: widget.onStartVoiceInput,
                       onStop: widget.onStopVoiceInput,
                     ),
@@ -240,7 +240,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.album,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsAlbum(),
+            icon: ImageUtil.toolsAlbum(),
             onTap: onTapAlbum,
             // onTap: () => PermissionUtil.storage(() => onTapAlbum?.call()),
           ),
@@ -249,7 +249,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.camera,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsCamera(),
+            icon: ImageUtil.toolsCamera(),
             onTap: onTapCamera,
             // onTap: () => PermissionUtil.camera(() => onTapCamera?.call()),
           ),
@@ -258,7 +258,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.videoCall,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsVideoCall(),
+            icon: ImageUtil.toolsVideoCall(),
             onTap: onTapVideoCall,
             // onTap: () => PermissionUtil.camera(() => onTapVideoCall?.call()),
           ),
@@ -267,7 +267,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.location,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsLocation(),
+            icon: ImageUtil.toolsLocation(),
             onTap: onTapLocation,
             // onTap: () => PermissionUtil.location(() => onTapLocation?.call()),
           ),
@@ -276,7 +276,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.file,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsFile(),
+            icon: ImageUtil.toolsFile(),
             onTap: onTapFile,
             // onTap: () => PermissionUtil.storage(() => onTapFile?.call()),
           ),
@@ -285,7 +285,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.carte,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsCarte(),
+            icon: ImageUtil.toolsCarte(),
             onTap: onTapCarte,
           ),
         ),
@@ -293,7 +293,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.voiceInput,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsVoiceInput(),
+            icon: ImageUtil.toolsVoiceInput(),
             onTap: onTapVoiceInput,
             // onTap: () => PermissionUtil.microphone(
             //   () => onTapVoiceInput?.call(),
@@ -389,7 +389,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
             child: Container(
               color: Colors.white,
               child: Center(
-                child: IconUtil.voiceInputNor(),
+                child: ImageUtil.voiceInputNor(),
               ),
             ),
           ),
@@ -407,7 +407,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                 label: UILocalizations.album,
                 style: toolsTextStyle,
                 image: _buildBtn(
-                  icon: IconUtil.toolsAlbum(),
+                  icon: ImageUtil.toolsAlbum(),
                   onTap: onTapAlbum,
                 ),
               )),
@@ -418,7 +418,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                 label: UILocalizations.file,
                 style: toolsTextStyle,
                 image: _buildBtn(
-                  icon: IconUtil.toolsFile(),
+                  icon: ImageUtil.toolsFile(),
                   onTap: onTapFile,
                 ),
               )),
@@ -431,7 +431,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                 label: UILocalizations.camera,
                 style: toolsTextStyle,
                 image: _buildBtn(
-                  icon: IconUtil.toolsCamera(),
+                  icon: ImageUtil.toolsCamera(),
                   onTap: onTapCamera,
                 ),
               )),
@@ -442,7 +442,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                 label: UILocalizations.carte,
                 style: toolsTextStyle,
                 image: _buildBtn(
-                  icon: IconUtil.toolsCarte(),
+                  icon: ImageUtil.toolsCarte(),
                   onTap: onTapCarte,
                 ),
               )),
@@ -455,7 +455,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                 label: UILocalizations.videoCall,
                 style: toolsTextStyle,
                 image: _buildBtn(
-                  icon: IconUtil.toolsVideoCall(),
+                  icon: ImageUtil.toolsVideoCall(),
                   onTap: onTapVideoCall,
                 ),
               )),
@@ -466,7 +466,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
                 label: UILocalizations.voiceInput,
                 style: toolsTextStyle,
                 image: _buildBtn(
-                  icon: IconUtil.toolsVoiceInput(),
+                  icon: ImageUtil.toolsVoiceInput(),
                   onTap: onTapVoiceInput,
                 ),
               )),
@@ -477,7 +477,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
             label: UILocalizations.location,
             style: toolsTextStyle,
             image: _buildBtn(
-              icon: IconUtil.toolsLocation(),
+              icon: ImageUtil.toolsLocation(),
               onTap: onTapLocation,
             ),
           )),
@@ -489,7 +489,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.album,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsAlbum(),
+            icon: ImageUtil.toolsAlbum(),
             onTap: onTapAlbum,
             // onTap: () => PermissionUtil.storage(() => onTapAlbum?.call()),
           ),
@@ -498,7 +498,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.camera,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsCamera(),
+            icon: ImageUtil.toolsCamera(),
             onTap: onTapCamera,
             // onTap: () => PermissionUtil.camera(() => onTapCamera?.call()),
           ),
@@ -507,7 +507,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.videoCall,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsVideoCall(),
+            icon: ImageUtil.toolsVideoCall(),
             onTap: onTapVideoCall,
             // onTap: () => PermissionUtil.camera(() => onTapVideoCall?.call()),
           ),
@@ -516,7 +516,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.location,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsLocation(),
+            icon: ImageUtil.toolsLocation(),
             onTap: onTapLocation,
             // onTap: () => PermissionUtil.location(() => onTapLocation?.call()),
           ),
@@ -525,7 +525,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.file,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsFile(),
+            icon: ImageUtil.toolsFile(),
             onTap: onTapFile,
             // onTap: () => PermissionUtil.storage(() => onTapFile?.call()),
           ),
@@ -534,7 +534,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.carte,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsCarte(),
+            icon: ImageUtil.toolsCarte(),
             onTap: onTapCarte,
           ),
         ),
@@ -542,7 +542,7 @@ class _ChatToolsViewState extends State<ChatToolsView>
           label: UILocalizations.voiceInput,
           style: toolsTextStyle,
           image: _buildBtn(
-            icon: IconUtil.toolsVoiceInput(),
+            icon: ImageUtil.toolsVoiceInput(),
             onTap: onTapVoiceInput,
             // onTap: () => PermissionUtil.microphone(
             //   () => onTapVoiceInput?.call(),
