@@ -4,9 +4,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_openim_widget/flutter_openim_widget.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import 'chat_voice_record_bar.dart';
-import 'chat_voice_record_view.dart';
-
 typedef SpeakViewChildBuilder = Widget Function(ChatVoiceRecordBar recordBar);
 
 class ChatVoiceRecordLayout extends StatefulWidget {
@@ -39,12 +36,11 @@ class _ChatVoiceRecordLayoutState extends State<ChatVoiceRecordLayout> {
 
   @override
   void initState() {
-    UILocalizations.setLocale(widget.locale);
+    UILocalizations.set(widget.locale);
     super.initState();
   }
 
   void callback(int sec, String path) {
-    print('----path:$path------sec:$sec');
     _sec = sec;
     _path = path;
   }
@@ -70,7 +66,7 @@ class _ChatVoiceRecordLayoutState extends State<ChatVoiceRecordLayout> {
         },
         onLongPressEnd: (details) async {
           await _record.stop();
-          print('---------停止记录--------------');
+          // 停止记录
           setState(() {
             if (_selectedPressArea) {
               _callback();
@@ -101,7 +97,7 @@ class _ChatVoiceRecordLayoutState extends State<ChatVoiceRecordLayout> {
         },
         onLongPressStart: (details) {
           setState(() {
-            print('---------开始记录--------------');
+            // 开始记录
             _record = VoiceRecord(callback);
             _record.start();
             _selectedPressArea = true;
@@ -154,7 +150,6 @@ class _ChatVoiceRecordLayoutState extends State<ChatVoiceRecordLayout> {
   }
 
   void _callback() {
-    print('------------发送------------');
     if (_sec > 0 && null != _path) {
       widget.onCompleted?.call(_sec, _path!);
     }
