@@ -1,10 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_openim_widget/flutter_openim_widget.dart';
-import 'package:flutter_openim_widget/src/unread_count_view.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:flutter_slidable/flutter_slidable.dart';
-
-import 'chat_avatar_view.dart';
 
 final pinColors = [Color(0xFF87C0FF), Color(0xFF0060E7)];
 final deleteColors = [Color(0xFFFFC84C), Color(0xFFFFA93C)];
@@ -34,6 +30,8 @@ class ConversationItemView extends StatelessWidget {
   final List<MatchPattern> patterns;
   final Function()? onTap;
   final bool notDisturb;
+  final double extentRatio;
+  final bool isUserGroup;
 
   // final bool isPinned;
 
@@ -59,6 +57,8 @@ class ConversationItemView extends StatelessWidget {
     this.patterns = const [],
     this.onTap,
     this.notDisturb = false,
+    this.extentRatio = 0.5,
+    this.isUserGroup = false,
     // this.isPinned = false,
     this.titleStyle = const TextStyle(
       fontSize: 16,
@@ -103,10 +103,11 @@ class ConversationItemView extends StatelessWidget {
         timeStyle: timeStyle,
         onTap: onTap,
         notDisturb: notDisturb,
+        isUserGroup: isUserGroup,
       ),
       endActionPane: ActionPane(
         motion: DrawerMotion(),
-        // extentRatio: 0.75,
+        extentRatio: extentRatio,
         children: slideActions.map((e) => _SlidableAction(item: e)).toList(),
       ),
     );
@@ -139,6 +140,7 @@ class _ConversationView extends StatelessWidget {
     this.contentPrefixStyle,
     this.onTap,
     this.notDisturb = false,
+    this.isUserGroup = false,
   }) : super(key: key);
   final double avatarSize;
   final String? avatarUrl;
@@ -162,6 +164,7 @@ class _ConversationView extends StatelessWidget {
   final List<MatchPattern> patterns;
   final Function()? onTap;
   final bool notDisturb;
+  final bool isUserGroup;
 
   @override
   Widget build(BuildContext context) {
@@ -179,8 +182,9 @@ class _ConversationView extends StatelessWidget {
                 ChatAvatarView(
                   size: avatarSize,
                   url: avatarUrl,
-                  isCircle: isCircleAvatar ?? true,
+                  isCircle: isCircleAvatar ?? false,
                   borderRadius: avatarBorderRadius,
+                  isUserGroup: isUserGroup,
                 ),
                 SizedBox(width: 12.w),
                 Flexible(
@@ -237,7 +241,7 @@ class _ConversationView extends StatelessWidget {
                             Spacer(),
                             if (!notDisturb)
                               UnreadCountView(count: unreadCount),
-                            if (notDisturb) IconUtil.notDisturb(),
+                            if (notDisturb) ImageUtil.notDisturb(),
                           ],
                         ),
                       ],
@@ -278,6 +282,7 @@ class _SlidableAction extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Expanded(
+      flex: item.flex,
       child: GestureDetector(
         onTap: () {
           item.onTap?.call();
@@ -302,7 +307,7 @@ class _SlidableAction extends StatelessWidget {
           ),
           child: Container(
             alignment: Alignment.center,
-            width: item.width,
+            // width: item.width,
             padding: EdgeInsets.symmetric(horizontal: 12.w),
             child: Text(
               item.text,
@@ -312,6 +317,16 @@ class _SlidableAction extends StatelessWidget {
         ),
       ),
     );
+  }
+
+  // Here it is!
+  Size _textSize(String text, TextStyle style) {
+    final TextPainter textPainter = TextPainter(
+        text: TextSpan(text: text, style: style),
+        maxLines: 1,
+        textDirection: TextDirection.ltr)
+      ..layout(minWidth: 0, maxWidth: double.infinity);
+    return textPainter.size;
   }
 }
 
