@@ -5,14 +5,16 @@ import 'package:flutter_openim_widget/src/chat_emoji_view.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-class IconUtil {
-  IconUtil._();
+class ImageUtil {
+  ImageUtil._();
+
+  static final _package = "flutter_openim_widget";
 
   static String imageResStr(var name) => "assets/images/$name.webp";
 
   static AssetImage emojiImage(String key) => AssetImage(
-        IconUtil.imageResStr(emojiFaces[key]),
-        package: 'flutter_openim_widget',
+        ImageUtil.imageResStr(emojiFaces[key]),
+        package: _package,
       );
 
   static Widget svg(
@@ -27,7 +29,7 @@ class IconUtil {
       width: width,
       height: height,
       color: color,
-      package: 'flutter_openim_widget',
+      package: _package,
     );
   }
 
@@ -44,9 +46,8 @@ class IconUtil {
       height: height,
       fit: fit,
       color: color,
-      // cacheHeight: height?.toInt(),
       // cacheWidth: width?.toInt(),
-      package: 'flutter_openim_widget',
+      package: _package,
     );
   }
 
@@ -149,52 +150,55 @@ class IconUtil {
         height: 48.h,
       );
 
-  static Widget menuCopy() => assetImage(
-        'ic_menu_copy',
-        width: 18.w,
-        height: 18.h,
-      );
+  static Widget menuCopy() =>
+      assetImage('ic_menu_copy', width: 18.w, height: 18.w);
 
   static Widget menuDel() => assetImage(
         'ic_menu_del',
         width: 18.w,
-        height: 18.h,
+        height: 18.w,
       );
 
   static Widget menuForward() => assetImage(
         'ic_menu_forward',
-        width: 18.w,
-        height: 18.h,
+        width: 16.w,
+        height: 16.w,
       );
 
   static Widget menuMultiChoice() => assetImage(
         'ic_menu_multichoice',
         width: 18.w,
-        height: 18.h,
+        height: 18.w,
       );
 
   static Widget menuReply() => assetImage(
         'ic_menu_reply',
         width: 18.w,
-        height: 18.h,
+        height: 18.w,
       );
 
   static Widget menuRevoke() => assetImage(
         'ic_menu_revoke',
         width: 18.w,
-        height: 18.h,
+        height: 18.w,
       );
 
   static Widget menuDownload() => assetImage(
         'ic_menu_download',
         width: 18.w,
-        height: 18.h,
+        height: 18.w,
       );
 
   static Widget menuTranslation() => assetImage(
         'ic_menu_translation',
         width: 18.w,
-        height: 18.h,
+        height: 18.w,
+      );
+
+  static Widget menuAddEmoji() => assetImage(
+        'ic_menu_add_emoji',
+        width: 19.w,
+        height: 19.w,
       );
 
   static Widget file() => assetImage(
@@ -206,7 +210,7 @@ class IconUtil {
   static Widget delQuote() => assetImage(
         'ic_del_quote',
         width: 14.w,
-        height: 15.h,
+        height: 15.w,
       );
 
   static Widget voiceInputNor() => assetImage(
@@ -236,18 +240,25 @@ class IconUtil {
     required String url,
     double? width,
     double? height,
+    int? cacheWidth,
+    int? cacheHeight,
     BoxFit? fit,
     bool loadProgress = true,
+    bool clearMemoryCacheWhenDispose = true,
+    bool lowMemory = true,
+    Widget? errorWidget,
   }) =>
-      networkImage(
+      _extendedImage(
         url: url,
         width: width,
         height: height,
+        cacheWidth: cacheHeight,
+        cacheHeight: cacheHeight,
         fit: fit,
-        cacheWidth: width?.toInt(),
-        cacheHeight: height?.toInt(),
         loadProgress: loadProgress,
-        clearMemoryCacheWhenDispose: true,
+        clearMemoryCacheWhenDispose: clearMemoryCacheWhenDispose,
+        lowMemory: lowMemory,
+        errorWidget: errorWidget,
       );
 
   static Widget networkImage({
@@ -259,15 +270,42 @@ class IconUtil {
     BoxFit? fit,
     bool loadProgress = true,
     bool clearMemoryCacheWhenDispose = false,
+    bool lowMemory = true,
+    Widget? errorWidget,
+  }) =>
+      lowMemoryNetworkImage(
+        url: url,
+        width: width,
+        height: height,
+        cacheWidth: cacheWidth,
+        cacheHeight: cacheHeight,
+        fit: fit,
+        loadProgress: loadProgress,
+        clearMemoryCacheWhenDispose: clearMemoryCacheWhenDispose,
+        lowMemory: lowMemory,
+        errorWidget: errorWidget,
+      );
+
+  static Widget _extendedImage({
+    required String url,
+    double? width,
+    double? height,
+    int? cacheWidth,
+    int? cacheHeight,
+    BoxFit? fit,
+    bool loadProgress = true,
+    bool clearMemoryCacheWhenDispose = true,
+    bool lowMemory = true,
+    Widget? errorWidget,
   }) =>
       ExtendedImage.network(
         url,
         width: width,
         height: height,
         fit: fit,
+        cacheWidth: lowMemory ? cacheWidth ?? (1.sw * .75).toInt() : null,
+        cacheHeight: lowMemory ? cacheHeight : null,
         cache: true,
-        cacheWidth: cacheWidth ?? (1.sw * .75).toInt(),
-        cacheHeight: cacheHeight,
         clearMemoryCacheWhenDispose: clearMemoryCacheWhenDispose,
         loadStateChanged: (ExtendedImageState state) {
           switch (state.extendedImageLoadState) {
@@ -280,95 +318,62 @@ class IconUtil {
                             loadingProgress.expectedTotalBytes!
                         : null;
                 // CupertinoActivityIndicator()
-                return loadProgress
-                    ? Container(
-                        width: 15.0,
-                        height: 15.0,
-                        child: Center(
+                return Container(
+                  width: 15.0,
+                  height: 15.0,
+                  child: loadProgress
+                      ? Center(
                           child: CircularProgressIndicator(
                             strokeWidth: 1.5,
-                            value: progress,
+                            value: progress ?? 0,
                           ),
-                        ),
-                      )
-                    : Container();
+                        )
+                      : null,
+                );
               }
             case LoadState.completed:
-              {
-                ///if you don't want override completed widget
-                ///please return null or state.completedWidget
-                return null;
-                //return state.completedWidget;
-                // return state.completedWidget;
-                // return FadeIn(
-                //   // duration: const Duration(milliseconds: 100),
-                //   child: ExtendedRawImage(
-                //     image: state.extendedImageInfo?.image,
-                //     width: width,
-                //     height: height,
-                //     fit: fit,
-                //   ),
-                // );
-              }
+              return null;
             case LoadState.failed:
-              //remove memory cached
+              // remove memory cached
               state.imageProvider.evict();
-              return error(width: width, height: height);
+              return errorWidget ?? error(width: width, height: height);
           }
         },
-        // border: Border.all(color: Colors.red, width: 1.0),
-        // shape: boxShape,
-        // borderRadius: BorderRadius.all(Radius.circular(30.0)),
-        //cancelToken: cancellationToken,
       );
-/*FadeInImage.memoryNetwork(
-        placeholder: kTransparentImage,
-        image: url,
-        height: height,
-        width: width,
-        fit: fit ?? BoxFit.cover,
-        imageCacheHeight: cacheHeight,
-        imageCacheWidth: cacheWidth,
-        fadeInDuration: const Duration(milliseconds: 100),
-      );*/
-/*Image.network(
-        url,
-        height: height,
-        width: width,
-        fit: fit ?? BoxFit.cover,
-        cacheHeight: cacheHeight,
-        cacheWidth: cacheWidth,
-        loadingBuilder: null != loadingWidget
-            ? (context, child, loadingProgress) => loadingWidget
-            : null,
-        errorBuilder: null != errorWidget
-            ? (context, Object, stackTrace) => errorWidget
-            : null,
-      );*/
 
-/*static Widget networkImage({
+/*static Widget _cachedNetworkImage({
     required String url,
     double? width,
     double? height,
+    int? cacheWidth,
+    int? cacheHeight,
     BoxFit? fit,
-    int? memCacheHeight,
-    int? memCacheWidth,
-    PlaceholderWidgetBuilder? placeholder,
-    ProgressIndicatorBuilder? progressIndicatorBuilder,
-    LoadingErrorWidgetBuilder? errorWidget,
+    bool loadProgress = true,
+    bool clearMemoryCacheWhenDispose = true,
+    bool lowMemory = true,
+    Widget? errorWidget,
   }) =>
       CachedNetworkImage(
         imageUrl: url,
         width: width,
         height: height,
         fit: fit,
-        memCacheHeight: memCacheHeight,
-        memCacheWidth: memCacheWidth,
-        placeholder: placeholder,
-        progressIndicatorBuilder: progressIndicatorBuilder,
-        errorWidget: errorWidget,
-        // filterQuality: FilterQuality.medium,
-        // cacheManager: CustomCacheManager.instance,
+        memCacheWidth: cacheWidth ?? (lowMemory ? (1.sw * .75).toInt() : null),
+        memCacheHeight: cacheHeight,
+        // placeholder: placeholder,
+        progressIndicatorBuilder: (context, url, progress) => Container(
+          width: 10.0,
+          height: 10.0,
+          child: loadProgress
+              ? Center(
+                  child: CircularProgressIndicator(
+                    strokeWidth: 1.5,
+                    value: progress.progress ?? 0,
+                  ),
+                )
+              : null,
+        ),
+        errorWidget: (_, url, er) =>
+            errorWidget ?? error(width: width, height: height),
       );*/
-
 }
